# assets/swagger.json
{"swagger":"2.0","info":{"version":"1.0","title":"SlahsMobility Users and providers service","license":{"name":"Apache License Version 2.0"}},"host":"localhost:8080","tags":[{"name":"auth-controller","description":"Auth Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"product-controller","description":"Product Controller"},{"name":"provider-controller","description":"Provider Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/api/auth/authenticate":{"post":{"tags":["auth-controller"],"summary":"createToken","operationId":"createTokenUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"authenticationRequest","description":"authenticationRequest","required":true,"schema":{"$ref":"#/definitions/AuthenticationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AuthenticationResponse"}}}}},"/api/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}}}}},"/api/products/":{"get":{"tags":["product-controller"],"summary":"getProducts","operationId":"getProductsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}}}},"post":{"tags":["product-controller"],"summary":"createProduct","operationId":"createProductUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"product","description":"product","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/products/cities/{city}":{"get":{"tags":["product-controller"],"summary":"getProductsByCity","operationId":"getProductsByCityUsingGET","produces":["*/*"],"parameters":[{"name":"city","in":"path","description":"city","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}}}}},"/api/products/types/{type}":{"get":{"tags":["product-controller"],"summary":"getProductsByType","operationId":"getProductsByTypeUsingGET","produces":["*/*"],"parameters":[{"name":"type","in":"path","description":"type","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}}}}},"/api/products/{productId}":{"get":{"tags":["product-controller"],"summary":"getProduct","operationId":"getProductUsingGET","produces":["*/*"],"parameters":[{"name":"productId","in":"path","description":"productId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Product"}}}},"put":{"tags":["product-controller"],"summary":"updateProduct","operationId":"updateProductUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"productId","in":"path","description":"productId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"productUpdateRequest","description":"productUpdateRequest","required":true,"schema":{"$ref":"#/definitions/ProductUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/providers/":{"get":{"tags":["provider-controller"],"summary":"getProviders","operationId":"getProvidersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Provider"}}}}},"post":{"tags":["provider-controller"],"summary":"createProvider","operationId":"createProviderUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"provider","description":"provider","required":true,"schema":{"$ref":"#/definitions/Provider"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/providers/{providerId}":{"get":{"tags":["provider-controller"],"summary":"getProvider","operationId":"getProviderUsingGET","produces":["*/*"],"parameters":[{"name":"providerId","in":"path","description":"providerId","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Provider"}}}},"put":{"tags":["provider-controller"],"summary":"updateProvider","operationId":"updateProviderUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"providerId","in":"path","description":"providerId","required":true,"type":"string"},{"in":"body","name":"providerUpdateRequest","description":"providerUpdateRequest","required":true,"schema":{"$ref":"#/definitions/ProviderUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/users/":{"get":{"tags":["user-controller"],"summary":"getUsers","operationId":"getUsersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserResponse"}}}}},"post":{"tags":["user-controller"],"summary":"createUser","operationId":"createUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"user","description":"user","required":true,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}},"put":{"tags":["user-controller"],"summary":"updateUser","operationId":"updateUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userUpdateRequest","description":"userUpdateRequest","required":true,"schema":{"$ref":"#/definitions/UserUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/users/password/reset":{"put":{"tags":["user-controller"],"summary":"resetPassword","operationId":"resetPasswordUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"resetPassword","description":"resetPassword","required":true,"schema":{"$ref":"#/definitions/ResetPassword"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}}}}},"/api/users/{userName}":{"get":{"tags":["user-controller"],"summary":"getUser","operationId":"getUserUsingGET","produces":["*/*"],"parameters":[{"name":"userName","in":"path","description":"userName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserResponse"}}}}},"/api/users/{userName}/email/confirmation":{"get":{"tags":["user-controller"],"summary":"emailConfirmation","operationId":"emailConfirmationUsingGET","produces":["*/*"],"parameters":[{"name":"userName","in":"path","description":"userName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}}},"definitions":{"AuthenticationRequest":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"AuthenticationRequest"},"AuthenticationResponse":{"type":"object","properties":{"access_token":{"type":"string"},"expires":{"type":"integer","format":"int64"},"refresh_token":{"type":"string"},"token_type":{"type":"string"}},"title":"AuthenticationResponse"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Product":{"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"image":{"type":"string"},"name":{"type":"string"},"providerName":{"type":"string"},"type":{"type":"string"}},"title":"Product"},"ProductUpdateRequest":{"type":"object","properties":{"description":{"type":"string"},"image":{"type":"string"},"name":{"type":"string"},"providerName":{"type":"string"},"type":{"type":"string"}},"title":"ProductUpdateRequest"},"Provider":{"type":"object","properties":{"address":{"type":"string"},"city":{"type":"string"},"name":{"type":"string"},"telephone":{"type":"string"}},"title":"Provider"},"ProviderUpdateRequest":{"type":"object","properties":{"address":{"type":"string"},"city":{"type":"string"},"name":{"type":"string"},"telephone":{"type":"string"}},"title":"ProviderUpdateRequest"},"ResetPassword":{"type":"object","properties":{"newPassword":{"type":"string"},"oldPassword":{"type":"string"}},"title":"ResetPassword"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"},"User":{"type":"object","properties":{"email":{"type":"string"},"emailConfirm":{"type":"boolean"},"lastName":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"userName":{"type":"string"}},"title":"User"},"UserResponse":{"type":"object","properties":{"email":{"type":"string"},"lastName":{"type":"string"},"name":{"type":"string"},"userName":{"type":"string"}},"title":"UserResponse"},"UserUpdateRequest":{"type":"object","properties":{"email":{"type":"string"},"lastName":{"type":"string"},"name":{"type":"string"}},"title":"UserUpdateRequest"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}