# src/main/java/com/slashmobility/UserProoviders/infraestructure/type/ProviderUpdateRequest.java
package com.slashmobility.UserProoviders.infraestructure.type;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Data;
import lombok.NoArgsConstructor;

@JsonInclude(JsonInclude.Include.NON_NULL)
@Data
@NoArgsConstructor
public class ProviderUpdateRequest {

    private String name;

    private String address;

    private String telephone;

    private String city;
}